requirements.txt
facebook-sdk==3.1.0
pillow